1. base64 encode/decode 2. tools.c / http-header-content-type compile type warning fixed.

+++ libhttp/source/http-header-content-type.c
@@ -19,7 +19,7 @@
 
 int http_header_content_type(const char* field, struct http_header_content_type_t *v)
 {
-	int* len;
+	size_t* len;
 	const char* p1;
 	const char* p = field;
